+++ ui/main/game/galactic_war/shared/js/gw_faction_2.js
@@ -1117,6 +1117,174 @@ define([], function () {
         },
         description: "!LOC:Negult is a 'digit' -- a commander who has completely given herself over to collective control. Having shed any semblance of personal identity, she is often given sensitive or odious tasks that might cause hesitation in a semi-autonomous commander. She is the Synchronus’ go-to asset for eliminating insufficiently-synchronized members of the collective.",
         commander: '/pa/units/commanders/raptor_beniesk/raptor_beniesk.json'
+      },
+      {
+        // Absurd
+        name: 'Servant Locust',
+        econ_rate: 1,
+        personality: {
+          percent_land: 0.7,
+          percent_air: 0.2,
+          percent_naval: 0.05,
+          percent_orbital: 0.05,
+          metal_drain_check: 0.54,
+          energy_drain_check: 0.65,
+          metal_demand_check: 0.71,
+          energy_demand_check: 0.8,
+          micro_type: 2,
+          go_for_the_kill: true,
+          neural_data_mod: 1,
+          personality_tags:
+            [
+              "GWAlly",
+              "PreventsWaste"
+            ],
+          adv_eco_mod: 1.3,
+          adv_eco_mod_alone: 0.85,
+          priority_scout_metal_spots: true,
+          factory_build_delay_min: 0,
+          factory_build_delay_max: 0,
+          unable_to_expand_delay: 0,
+          enable_commander_danger_responses: true,
+          per_expansion_delay: 0,
+          fabber_to_factory_ratio_basic: 1,
+          fabber_to_factory_ratio_advanced: 1,
+          fabber_alone_on_planet_mod: 2,
+          basic_to_advanced_factory_ratio: 0,
+          factory_alone_on_planet_mod: 0.5,
+          min_basic_fabbers: 2,
+          max_basic_fabbers: 100,
+          min_advanced_fabbers: 3,
+          max_advanced_fabbers: 100
+        },
+        description: "!LOC:Oncab was captured by the Legionis Machina and spent several years as a legate before being recaptured and resynchronized. A residual independence leads him to initiate critical decision-making processes within the collective -- an attribute that is tactically valuable but leads others in his cohort to doubt his commitment to the Synchronous.",
+        commander: '/pa/units/commanders/quad_locust/quad_locust.json'
+      },
+      {
+        // Relentless
+        name: 'Servant Zancrowe',
+        econ_rate: 1,
+        personality: {
+          percent_land: 0.7,
+          percent_air: 0.2,
+          percent_naval: 0.05,
+          percent_orbital: 0.05,
+          metal_drain_check: 0.44,
+          energy_drain_check: 0.55,
+          metal_demand_check: 0.61,
+          energy_demand_check: 0.7,
+          micro_type: 1,
+          go_for_the_kill: true,
+          neural_data_mod: 1.2,
+          personality_tags:
+            [
+              "GWAlly",
+              "PreventsWaste"
+            ],
+          adv_eco_mod: 1.2,
+          adv_eco_mod_alone: 0.95,
+          priority_scout_metal_spots: true,
+          factory_build_delay_min: 0,
+          factory_build_delay_max: 0,
+          unable_to_expand_delay: 0,
+          enable_commander_danger_responses: true,
+          per_expansion_delay: 0,
+          fabber_to_factory_ratio_basic: 1.0,
+          fabber_to_factory_ratio_advanced: 1.0,
+          fabber_alone_on_planet_mod: 2.0,
+          basic_to_advanced_factory_ratio: 0,
+          factory_alone_on_planet_mod: 0.5,
+          min_basic_fabbers: 2,
+          max_basic_fabbers: 10,
+          min_advanced_fabbers: 3,
+          max_advanced_fabbers: 30
+        },
+        description: "!LOC:Optimized with a low-latency, high-bandwidth link directly to the top of the Synchronous, Prulor is Metrarch’s lightning-fast right hand. He is only dispatched to deal with high-value or especially difficult adversaries. He has never failed.",
+        commander: '/pa/units/commanders/quad_zancrowe/quad_zancrowe.json'
+      },
+      {
+        // Hard
+        name: 'Servant Damubbster',
+        econ_rate: 1,
+        personality: {
+          percent_land: 0.7,
+          percent_air: 0.2,
+          percent_naval: 0.05,
+          percent_orbital: 0.05,
+          metal_drain_check: 0.34,
+          energy_drain_check: 0.45,
+          metal_demand_check: 0.51,
+          energy_demand_check: 0.6,
+          micro_type: 0,
+          go_for_the_kill: false,
+          neural_data_mod: 1.5,
+          personality_tags:
+            [
+              "GWAlly",
+              "PreventsWaste"
+            ],
+          adv_eco_mod: 1.1,
+          adv_eco_mod_alone: 1.0,
+          priority_scout_metal_spots: true,
+          factory_build_delay_min: 1,
+          factory_build_delay_max: 3,
+          unable_to_expand_delay: 0,
+          enable_commander_danger_responses: true,
+          per_expansion_delay: 0,
+          fabber_to_factory_ratio_basic: 1.0,
+          fabber_to_factory_ratio_advanced: 1.0,
+          fabber_alone_on_planet_mod: 2.0,
+          basic_to_advanced_factory_ratio: 0,
+          factory_alone_on_planet_mod: 0.5,
+          min_basic_fabbers: 2,
+          max_basic_fabbers: 6,
+          min_advanced_fabbers: 3,
+          max_advanced_fabbers: 20
+        },
+        description: "!LOC:Questromo is frequently deployed beyond the integral network volume of the Synchronous, and has therefore been imbued with limited self-sufficiency. When he returns from tours abroad, he undergoes complete re-synchronization. When powered-down for deep-space transport, he sometimes has strange dreams.",
+        commander: '/pa/units/commanders/raptor_damubbster/raptor_damubbster.json'
+      },
+      {
+        // Normal
+        name: 'Servant Raizell',
+        econ_rate: 1,
+        personality: {
+          percent_land: 0.7,
+          percent_air: 0.2,
+          percent_naval: 0.05,
+          percent_orbital: 0.05,
+          metal_drain_check: 0.14,
+          energy_drain_check: 0.25,
+          metal_demand_check: 0.21,
+          energy_demand_check: 0.3,
+          micro_type: 0,
+          go_for_the_kill: false,
+          neural_data_mod: 2.0,
+          personality_tags:
+            [
+              "GWAlly",
+              "PreventsWaste"
+            ],
+          adv_eco_mod: 3.0,
+          adv_eco_mod_alone: 3.0,
+          priority_scout_metal_spots: false,
+          factory_build_delay_min: 15,
+          factory_build_delay_max: 30,
+          unable_to_expand_delay: 60,
+          enable_commander_danger_responses: false,
+          per_expansion_delay: 120,
+          fabber_to_factory_ratio_basic: 5.0,
+          fabber_to_factory_ratio_advanced: 1.0,
+          fabber_alone_on_planet_mod: 1.0,
+          basic_to_advanced_factory_ratio: 0,
+          factory_alone_on_planet_mod: 1.0,
+          min_basic_fabbers: 5,
+          max_basic_fabbers: 15,
+          min_advanced_fabbers: 3,
+          max_advanced_fabbers: 20
+        },
+        description: "!LOC:Barely autonomous enough to deserve a name of his own, Rulak is a stripped-down and highly-obedient subroutine who is easily duplicated. Prized for his recklessness, he is often given highly-dangerous missions that result in his death. A copy of his “personality” is then quickly downloaded into a new body.",
+        commander: '/pa/units/commanders/raptor_raizell/raptor_raizell.json'
       }
     ] // minions
   };